requirements-day3.txt
fastapi==0.110.*
uvicorn[standard]==0.29.*
pydantic==2.*
python-dotenv==1.*
chromadb==0.4.*
google-generativeai==0.5.*
pypdf==4.*
streamlit==1.33.*
plotly==5.*
requests==2.*
